Update hacking_stuff.txt
I've been poking around the internets for information on getting an OSCP, and have added a bunch of things to my list. Now I'm realizing that the list is really difficult for me to go through, and doesn't necessarily have the information I want right where I need it and its easy for me to access. I'm going to think about how I want to re-work this so that I don't have to go between 20 different websites for all of my info, and so that its organized better.